@@ -10,7 +10,7 @@
     "aws-sdk": "^2.96.0",
     "babel-cli": "^6.24.1",
     "babel-core": "~6.24.1",
-    "babel-eslint": "^7.2.3",
+    "babel-eslint": "^8.2.2",
     "babel-helpers": "^6.24.1",
     "babel-jest": "~19.0.0",
     "babel-plugin-transform-async-to-generator": "^6.24.1",
@@ -23,7 +23,7 @@
     "cfn": "^1.2.0",
     "chai": "^4.1.1",
     "cucumber": "^2.3.1",
-    "eslint": "~3.19.0",
+    "eslint": "^4.10.0",
     "eslint-config-airbnb-base": "^11.3.1",
     "eslint-import-resolver-node": "^0.3.1",
     "eslint-plugin-flowtype": "^2.35.0",
@@ -32,7 +32,6 @@
     "fs": "0.0.1-security",
     "gulp": "^3.9.1",
     "gulp-clean": "^0.3.2",
-    "gulp-replace": "^0.6.1",
     "gulp-util": "^3.0.8",
     "gulp-zip": "^4.0.0",
     "husky": "^0.14.3",
@@ -43,7 +42,7 @@
     "request": "^2.81.0"
   },
   "dependencies": {
-    "awilix": "^2.7.0",
+    "awilix": "^3.0.7",
     "babel-runtime": "^6.25.0",
     "blipp": "^2.3.0",
     "global": "^4.3.2",
@@ -78,6 +77,8 @@
     "build": "babel src --out-dir dist --source-maps inline --copy-files",
     "prestart": "npm run build",
     "start": "node ./dist/server.js",
+    "prestartdebug": "npm run build",
+    "startdebug": "node --inspect=0.0.0.0:9229 ./dist/server.js",
     "predebug": "npm run build",
     "debug": "nodemon --inspect=0.0.0.0:9229 ./dist/server.js",
     "aat": "./node_modules/cucumber/bin/cucumber.js",
