@@ -1,6 +1,6 @@
 const hapiAuthJwt = require('hapi-auth-jwt2');
 
-export function register(server, options, next) {
+export function register(server: Object, options: Object, next: () => mixed) {
   const validate = (decoded, request, callback) => {
     // This validate function is meant to add any layers of security
     // we think are necessary. In most cases, no additional checks are
@@ -16,18 +16,16 @@ export function register(server, options, next) {
     return callback(null, true);
   };
 
-
   server.register(hapiAuthJwt, (err) => {
     if (err) {
       throw new Error(err);
     }
 
-    server.auth.strategy('jwt', 'jwt',
-      {
-        key: process.env.AUTH_SHARED_KEY || 'NeverShareYourSecret',
-        validateFunc: validate,
-        verifyOptions: { algorithms: ['HS256'] },
-      });
+    server.auth.strategy('jwt', 'jwt', {
+      key: process.env.AUTH_SHARED_KEY || 'NeverShareYourSecret',
+      validateFunc: validate,
+      verifyOptions: { algorithms: ['HS256'] },
+    });
 
     server.auth.default('jwt');
     next();
