@@ -1,7 +1,9 @@
-import type Encounter from '../../domain/types/encounter';
+import type { Encounter } from '../../domain/types/encounter';
+import type { Repository, Model, Error } from '../../domain/types/common';
 
-export default class Encounters {
-  constructor({ encounterModel }) {
+export default class Encounters implements Repository<number, Encounter> {
+  model: Model<number, Encounter>;
+  constructor({ encounterModel }: Object) {
     this.model = encounterModel;
   }
 
@@ -17,15 +19,15 @@ export default class Encounters {
     return this.model.create(encounter);
   }
 
-  update(id: number, encounter: Encounter): Promise<Encounter> {
+  update(id: number, encounter: Encounter): Promise<any> {
     return this.model
       .update(encounter, {
         where: { id },
         returning: true,
         plain: true,
       })
-      .then((result) => result && result[1])
-      .catch(() => ({ error: 'Encounter not found' }));
+      .then((result: any) => ((result && result[1] ? result[1] : { error: 'Encounter not found' }): Error))
+      .catch(() => ({ error: 'Encounter not found' }: Error));
   }
 
   delete(id: number): Promise<number> {
